+++ tests/test_crud_array.py
@@ -5,12 +5,12 @@ import pytest
 from playhouse.postgres_ext import ArrayField
 from pydantic import Field
 
-from pycurd.crud.ext.peewee_crud import PeeweeCrud
-from pycurd.crud.query_result_row import QueryResultRow
-from pycurd.crud.sql_crud import PlaceHolderGenerator, SQLExecuteResult
-from pycurd.query import QueryInfo
-from pycurd.types import RecordMapping
-from pycurd.values import ValuesToWrite
+from pycrud.crud.ext.peewee_crud import PeeweeCrud
+from pycrud.crud.query_result_row import QueryResultRow
+from pycrud.crud.sql_crud import PlaceHolderGenerator, SQLExecuteResult
+from pycrud.query import QueryInfo
+from pycrud.types import RecordMapping
+from pycrud.values import ValuesToWrite
 
 pytestmark = [pytest.mark.asyncio]
 
@@ -23,7 +23,7 @@ class TableOne(RecordMapping):
 class SpecialCrud(PeeweeCrud):
     async def execute_sql(self, sql: str, phg: PlaceHolderGenerator):
         self.last_sql = sql
-        return SQLExecuteResult(None, [])
+        return SQLExecuteResult(None, [[1],])
 
 
 def crud_db_init():
@@ -49,25 +49,34 @@ def crud_db_init():
     return c, db, TableOneModel
 
 
-async def test_curd_array_extend():
+async def test_crud_array_extend():
     c, db, TableOneModel = crud_db_init()
     await c.update(QueryInfo.from_table_raw(TableOne), values=ValuesToWrite({
         'arr.array_extend': ['aa', 'bb']
     }, table=TableOne, try_parse=True))
-    assert c.last_sql == 'UPDATE "table_one" SET "arr"="arr"||? WHERE "id" IN ()'
+    assert c.last_sql == 'UPDATE "table_one" SET "arr"="arr"||? WHERE "id" IN (?)'
 
 
-async def test_curd_array_extend_distinct():
+async def test_crud_array_extend_distinct():
     c, db, TableOneModel = crud_db_init()
     await c.update(QueryInfo.from_table_raw(TableOne), values=ValuesToWrite({
         'arr.array_extend_distinct': ['aa', 'bb']
     }, table=TableOne, try_parse=True))
-    assert c.last_sql == 'UPDATE "table_one" SET "arr"=ARRAY(SELECT DISTINCT unnest("arr"||?)) WHERE "id" IN ()'
+    assert c.last_sql == 'UPDATE "table_one" SET "arr"=ARRAY(SELECT DISTINCT unnest("arr"||?)) WHERE "id" IN (?)'
 
 
-async def test_curd_array_prune_distinct():
+async def test_crud_array_prune_distinct():
     c, db, TableOneModel = crud_db_init()
     await c.update(QueryInfo.from_table_raw(TableOne), values=ValuesToWrite({
         'arr.array_prune_distinct': ['aa', 'bb']
     }, table=TableOne, try_parse=True))
-    assert c.last_sql == 'UPDATE "table_one" SET "arr"=array(SELECT unnest("arr") EXCEPT SELECT unnest(?)) WHERE "id" IN ()'
+    assert c.last_sql == 'UPDATE "table_one" SET "arr"=array(SELECT unnest("arr") EXCEPT SELECT unnest(?)) WHERE "id" IN (?)'
+
+
+async def test_crud_array_contains_any():
+    c, db, TableOneModel = crud_db_init()
+    await c.get_list(QueryInfo.from_json(TableOne, {
+        '$select': 'id',
+        'arr.contains_any': ['a']
+    }))
+    assert c.last_sql == 'SELECT "id","id" FROM "table_one" WHERE "arr"&&? LIMIT 20'