+++ tests/test_query.py
@@ -2,9 +2,9 @@ import binascii
 import json
 from typing import Any, Union, List
 
-from pycurd.pydantic_ext.hex_string import HexString
-from pycurd.query import QueryInfo
-from pycurd.types import RecordMapping, RecordMappingField
+from pycrud.pydantic_ext.hex_string import HexString
+from pycrud.query import QueryInfo
+from pycrud.types import RecordMapping, RecordMappingField
 
 
 class User(RecordMapping):
@@ -17,7 +17,7 @@ class User(RecordMapping):
 
 def test_query_select_exclude():
     q = QueryInfo.from_table_raw(User, select=[User.id, User.nickname, User.password], select_exclude=[User.nickname])
-    assert q.select_for_curd == [User.id, User.password]
+    assert q.select_for_crud == [User.id, User.password]
 
 
 def test_values_bug_2():